Add `Object.refresh()` and fix PHP glitch with latest Haxe.
`Object.refresh()` clears the cache of related objects, so when you next call
the getter they will be loaded from database.  In future I may also get it to
update the properties of the current object from the database.

PHP glitch: The Haxe DB Macros (not the ufront DB macros) on platforms
other than neko, look for a field called `Manager`, and if it's not there
it creates a `__getManager()` field. Unfortunately this breaks when
`ufront.db.Object` doesn't have a maanger, but all of it's children do,
because of the build macro.  The `__getManager()` field is used when you
call `new MyModel()`, but in our case it ended up calling specifically
`ufront.db.Object.__getManager()`, which returned `Manager<Object>`, not
`Manager<T>`, and broke the inserting of objects.