Changed `id` to UId, not SUId. Clean up ufront.db.Object, better docs.
The `id` change is to add compatibility with SQLite.
Unless you already have over 2 billion rows in one table it shouldn't be an issue.

Get rid of `checkAuthRead` and `checkAuthWrite` which are not really
used, and hard to implement correctly given ufront's auth system,
dependency injection etc.

We might look at reintroducing later with a more complete implementation.

Removed ReadOnlyList, we're not using it any more.