Workaround for PHP metadata not attaching correctly.
Still not sure why Haxe does this, but if you don't access `Relationship.manager`
explicitly, but instead use `new Manager(relationship)`, then Haxe/PHP does not
attach the metadata to the Relationship model, and causes runtime exceptions.

This workaround fixes it for ManyToMany, which is the only use of this (unusual)
way of creating a Manager in the ufront-orm library.

Fixes #9