Fix serialization issues when using compiler cache
Basically an __cache__ field was being added, and also
after the first run a `posts:ManyToMany<A,B>` was creating
an `_posts:ManyToMany<A,B>` which was also being added.

We now explicitly ignore the cache field.
And we don't create an `_posts` field, we just use `posts` as is.