Changes to custom serialization.
Please note this changes the serialization format - so if you have serialized objects
in a persistant location (eg, stored in a user's session) these will not unserialize correctly.

You'll need to re-serialize them into the new format. So be careful!

But the new behaviour has some advantages:

- The macro still sets the default fields to serialize, and is consistent across platforms.
- You can optionally change the `hxSerializationFields` array before serialization to include other properties (like related objects)
- ManyToMany by default will include all the related objects (one `:A` and a `:List<B>`)
- If `manyToMany.useCompactSerialization` is set to true, it will only serialize the IDs (as well as any unsaved objects), for a smaller transfer size if you're only interested in *which* objects are joined, rather than the contents of every joined object.
- The serialization fields are now public and have API documentation... so it's less dark magic.

In future I'll attempt a macro like...

    return BlogPost.manager.all().alsoWithFields([ author, author.name, author.id, tags[].name ]);

...that is a type safe way of modifying `hxSerializationFields` on the relevant objects so you can easily pick which fields to return in your API.

Please note this is the start of a solution to https://github.com/ufront/ufront-mvc/issues/18