QueryBuilder should be called on a Connection rather than the model
Instead of

    Person.select()

I'm going for

    cnx.select(Person)

The reason is that I want to move away from `cnx` being the
`sys.db.Manager.cnx` singleton. I think we should be able
to inject the connection and use it wherever.

This will be important in an environment like NodeJS where you
can have more than 1 request, and possibly more than 1 connection,
active at a time.